# docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Address:
    properties:
      city:
        description: City name
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      number:
        description: Building number
        type: string
      owner_id:
        description: Owner's ID
        type: integer
      owner_type:
        description: Owner's type
        type: string
      postal:
        description: Postal code
        type: string
      street:
        description: Street name
        type: string
      updatedAt:
        type: string
    type: object
...
  model.Grade:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        description: Name is the name of the grade.
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.Score:
    properties:
      comment:
        description: Comment represents any additional comments about the score.
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      exam_id:
        description: ExamID represents the ID of the associated exam.
        type: integer
      grade:
        allOf:
        - $ref: '#/definitions/model.Grade'
        description: Grade represents the associated grade.
      grade_id:
        description: GradeID represents the ID of the associated grade.
        type: integer
      id:
        type: integer
      points:
        description: Points represents the number of points scored by the student.
        type: integer
      student_id:
        description: StudentID represents the ID of the student.
        type: integer
      updatedAt:
        type: string
    type: object
  model.Student:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/model.Address'
        description: Address represents the address of the student
      class_id:
        description: ClassID represents the ID of the class the student belongs to
        type: integer
      createdAt:
        type: string
      date_of_birth:
        description: DateOfBirth represents the date of birth of the student
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nachname:
        description: Nachname represents the last name of the student
        type: string
      updatedAt:
        type: string
      vorname:
        description: Vorname represents the first name of the student
        type: string
    type: object
...
  routes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:1323
info:
  contact:
    email: patrykhegenberg@gmail.com
    name: API support
    url: http://swagger.io/support
  license:
    name: Apache 2.0
    url: http://apache.com/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: Notenverwaltung API
  version: 0.0.8
paths:
...
  /scores:
    get:
      consumes:
      - application/json
      description: get all scores from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Score'
            type: array
        "500":
          description: Fehler beim Abrufen der Scores
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all scores
      tags:
      - score
    post:
      consumes:
      - application/json
      description: create a new score
      parameters:
      - description: Score object that needs to be added
        in: body
        name: score
        required: true
        schema:
          $ref: '#/definitions/model.Score'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Score'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des Scores
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create score
      tags:
      - score
  /scores/:id:
    delete:
      consumes:
      - application/json
      description: delete an existing score
      parameters:
      - description: Score ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Ungültige Score-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Löschen des Scores
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: delete score
      tags:
      - score
    get:
      consumes:
      - application/json
      description: get one score from db by ID.
      parameters:
      - description: Score ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Score'
        "400":
          description: Ungültige Score-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Score nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get one score by id
      tags:
      - score
    put:
      consumes:
      - application/json
      description: update an existing score
      parameters:
      - description: Score ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated score object
        in: body
        name: score
        required: true
        schema:
          $ref: '#/definitions/model.Score'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Score'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Score nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: update score
      tags:
      - score
  /students:
    get:
      consumes:
      - application/json
      description: get all students from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Student'
            type: array
        "500":
          description: Fehler beim Abrufen der Studenten
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all students
      tags:
      - student
    post:
      consumes:
      - application/json
      description: create a new student
      parameters:
      - description: Student object that needs to be added
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/model.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Student'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des Studenten
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create student
      tags:
      - student
  /students/:id:
    delete:
      consumes:
      - application/json
      description: delete an existing student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Ungültige Studenten-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Löschen des Studenten
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: delete student
      tags:
      - student
    get:
      consumes:
      - application/json
      description: get one student from db by ID.
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Student'
        "400":
          description: Ungültige Studenten-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Student nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get one student by id
      tags:
      - student
    put:
      consumes:
      - application/json
      description: update an existing student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated student object
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/model.Student'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Student'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Student nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: update student
      tags:
      - student
  /students/class/:id:
    get:
      consumes:
      - application/json
      description: get all students from db by class ID.
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Student'
            type: array
        "400":
          description: Ungültige Class-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Keine Students gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all students by a specific class id
      tags:
      - student
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"